+++ packages/builder/builder-rspack-provider/src/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './fs';
 export * from './constants';
-export * from './getRspackVersion';
+export * from './rspackVersion';