+++ packages/builder/builder-webpack-provider/src/index.ts
@@ -38,5 +38,4 @@ export type {
   WebpackConfig,
   CSSExtractOptions,
   HTMLPluginOptions,
-  TerserPluginOptions,
 } from './types';