+++ packages/generator/plugins/generator-plugin/modern.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   buildConfig: {
+    target: 'es2019',
     autoExternal: false,
     alias: {
       chalk: '@modern-js/utils/chalk',