+++ packages/solutions/app-tools/src/analyze/nestedRoutes.ts
@@ -78,11 +78,14 @@ export const optimizeRoute = (
   }
 
   const { children } = routeTree;
+  const hasPage = children.some(child => child.index);
   if (
     !routeTree._component &&
     !routeTree.error &&
     !routeTree.loading &&
-    !routeTree.config
+    !routeTree.config &&
+    !routeTree.clientData &&
+    !hasPage
   ) {
     const newRoutes = children.map(child => {
       const routePath = `${routeTree.path ? routeTree.path : ''}${
@@ -102,7 +105,6 @@ export const optimizeRoute = (
       }
       return newRoute;
     });
-
     return Array.from(new Set(newRoutes)).flatMap(optimizeRoute);
   } else {
     const optimizedChildren = routeTree.children.flatMap(optimizeRoute);
@@ -191,6 +193,14 @@ export const walk = async (
       }
     }
 
+    if (itemWithoutExt === NESTED_ROUTE.LAYOUT_CLIENT_LOADER) {
+      route.clientData = itemPath;
+    }
+
+    if (itemWithoutExt === NESTED_ROUTE.LAYOUT_DATA_FILE) {
+      route.data = itemPath;
+    }
+
     if (itemWithoutExt === NESTED_ROUTE.LAYOUT_CONFIG_FILE) {
       if (!route.config) {
         route.config = itemPath;
@@ -203,6 +213,14 @@ export const walk = async (
 
     if (itemWithoutExt === NESTED_ROUTE.PAGE_LOADER_FILE) {
       pageLoaderFile = itemPath;
+    }
+
+    if (itemWithoutExt === NESTED_ROUTE.PAGE_CLIENT_LOADER) {
+      route.clientData = itemPath;
+    }
+
+    if (itemWithoutExt === NESTED_ROUTE.PAGE_DATA_FILE) {
+      route.data = itemPath;
     }
 
     if (itemWithoutExt === NESTED_ROUTE.PAGE_CONFIG_FILE) {