+++ tests/e2e/builder/cases/styled-component/index.swc.test.ts
@@ -14,6 +14,10 @@ const commonConfig = {
     },
     output: {
       disableMinimize: true,
+      // avoid unwrapOutputJSON conflict
+      distPath: {
+        root: 'dist-swc',
+      },
     },
   },
 };