+++ tests/integration/module/src/fixtures/transformLodash/src/a.js
@@ -1,0 +1,5 @@
+import _ from 'lodash';
+
+export const libs = {
+  lodash: _,
+};