+++ tests/integration/ssr/fixtures/preload/modern.config.ts
@@ -1,0 +1,14 @@
+import { appTools, defineConfig } from '@modern-js/app-tools';
+
+export default defineConfig({
+  server: {
+    ssr: {
+      preload: true,
+    },
+  },
+  runtime: {
+    router: true,
+  },
+
+  plugins: [appTools()],
+});