+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @modern-js/generator-utils
 
+## 3.1.35
+
+### Patch Changes
+
+- Updated dependencies [15d30abdc66]
+  - @modern-js/utils@2.31.2
+  - @modern-js/plugin-i18n@2.31.2
+  - @modern-js/generator-common@3.1.35
+
+## 3.1.34
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.31.1
+- @modern-js/utils@2.31.1
+- @modern-js/generator-common@3.1.34
+
+## 3.1.33
+
+### Patch Changes
+
+- Updated dependencies [1882366]
+  - @modern-js/utils@2.31.0
+  - @modern-js/plugin-i18n@2.31.0
+  - @modern-js/generator-common@3.1.33
+
 ## 3.1.32
 
 ### Patch Changes