+++ packages/server/plugin-express/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @modern-js/plugin-express
 
+## 2.31.2
+
+### Patch Changes
+
+- Updated dependencies [15d30abdc66]
+  - @modern-js/utils@2.31.2
+  - @modern-js/bff-core@2.31.2
+  - @modern-js/bff-runtime@2.31.2
+  - @modern-js/types@2.31.2
+
+## 2.31.1
+
+### Patch Changes
+
+- @modern-js/bff-core@2.31.1
+- @modern-js/bff-runtime@2.31.1
+- @modern-js/types@2.31.1
+- @modern-js/utils@2.31.1
+
+## 2.31.0
+
+### Patch Changes
+
+- Updated dependencies [1882366]
+  - @modern-js/utils@2.31.0
+  - @modern-js/bff-core@2.31.0
+  - @modern-js/bff-runtime@2.31.0
+  - @modern-js/types@2.31.0
+
 ## 2.30.0
 
 ### Patch Changes